[#6]refactor: 재사용가능한 cell 생성
