+++ ProjectManager/ProjectManager/ViewControllers/TableViewControllers/TODOTableViewController.swift
@@ -13,7 +13,7 @@ class TODOTableViewController: UITableViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        tableView.register(UITableViewCell.self, forCellReuseIdentifier: "reuseIdentifier")
+        tableView.register(ScheduleCell.classForCoder(), forCellReuseIdentifier: "scheduleCell")
         self.tableView.separatorStyle = .none
         
         let header: UIView = {
@@ -92,64 +92,12 @@ class TODOTableViewController: UITableViewController {
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         
         let cell: UITableViewCell = {
-            let cell = tableView.dequeueReusableCell(withIdentifier: "reuseIdentifier", for: indexPath)
+            let cell = tableView.dequeueReusableCell(withIdentifier: "scheduleCell", for: indexPath)
             cell.backgroundColor = .magenta
 
             return cell
         }()
-        
-        let titleLabel: UILabel = {
-            let titleLabel = UILabel()
-            titleLabel.backgroundColor = .blue
-            titleLabel.text = "title"
-            titleLabel.lineBreakMode = .byTruncatingTail
-            titleLabel.translatesAutoresizingMaskIntoConstraints = false
-            
-            return titleLabel
-        }()
 
-        let descriptionLabel: UILabel = {
-            let descriptionLabel = UILabel()
-            descriptionLabel.backgroundColor = .yellow
-            descriptionLabel.text = """
-I'm on the next level yeah절대적 룰을 지켜내 손을놓지말아결속은 나의 무기광야로 걸어가알아 네 homeground위협에 맞서서제껴라 제껴라 제껴라상상도 못한 black out유혹은 깊고 진해(Too hot too hot)(Ooh ooh wee) 맞잡은 손을 놓쳐난 절대 포기 못해I'm on the next level저 너머의 문을 열어Next level널 결국엔 내가 부셔Next levelKosmo에 닿을 때까지Next level제껴라 제껴라 제껴라
-"""
-            descriptionLabel.numberOfLines = 3
-            descriptionLabel.lineBreakMode = .byTruncatingTail
-            descriptionLabel.textColor = .black
-            descriptionLabel.translatesAutoresizingMaskIntoConstraints = false
-
-            return descriptionLabel
-        }()
-
-        let dateLabel: UILabel = {
-            let dateLabel = UILabel()
-            dateLabel.backgroundColor = .red
-            dateLabel.text = "date"
-            dateLabel.translatesAutoresizingMaskIntoConstraints = false
-
-            return dateLabel
-        }()
-        
-        cell.addSubview(titleLabel)
-        cell.addSubview(descriptionLabel)
-        cell.addSubview(dateLabel)
-
-        NSLayoutConstraint.activate([
-            cell.topAnchor.constraint(equalTo: titleLabel.topAnchor),
-            cell.bottomAnchor.constraint(equalTo: dateLabel.bottomAnchor),
-        
-            titleLabel.topAnchor.constraint(equalTo: cell.topAnchor),
-            titleLabel.leadingAnchor.constraint(equalTo: cell.leadingAnchor),
-            titleLabel.trailingAnchor.constraint(equalTo: cell.trailingAnchor),
-            
-            descriptionLabel.widthAnchor.constraint(equalTo: cell.widthAnchor),
-            descriptionLabel.topAnchor.constraint(equalTo: titleLabel.bottomAnchor),
-            
-            dateLabel.widthAnchor.constraint(equalTo: cell.widthAnchor),
-            dateLabel.topAnchor.constraint(equalTo: descriptionLabel.bottomAnchor)
-        ])
-        
         return cell
     }
 }
